separate components into directories
