+++ src/components/search/knowledge-graphs/index.js
@@ -1,0 +1,1 @@
+export * from './knowledge-graphs'
