+++ src/components/search/pagination-tray/index.js
@@ -1,0 +1,1 @@
+export * from './pagination-tray'
