+++ src/components/search/pagination-tray/pagination-tray.css
