+++ src/components/search/result-card/index.js
@@ -1,0 +1,1 @@
+export * from './result-card'
