+++ src/components/search/result-card/result-card.css
