+++ src/components/search/results/index.js
@@ -1,0 +1,1 @@
+export * from './results'
